move item managers under repository utils
