+++ src/main/java/david/buzas/whattoeat/repositories/utils/ItemManager.java
@@ -1,4 +1,4 @@
-package david.buzas.whattoeat.utils.itemmanager;
+package david.buzas.whattoeat.repositories.utils;
 
 import java.io.IOException;
 import java.util.List;
