+++ src/main/java/david/buzas/whattoeat/repositories/utils/JsonFileItemManager.java
@@ -1,4 +1,4 @@
-package david.buzas.whattoeat.utils.itemmanager;
+package david.buzas.whattoeat.repositories.utils;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
